vwsfriend/mqtt_extra_requirements.txt
weconnect-mqtt~=0.42.1
